pyproject.toml
[tool.poetry]
name = "exchangereader"
version = "0.1.0"
description = ""
authors = ["James Groves <james.groves@shopify.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
black = "^25.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
